move meshconn to toplevel
