+++ examples/raft-on-mesh/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/coreos/etcd/raft/raftpb"
 
 	"github.com/weaveworks/mesh"
-	"github.com/weaveworks/mesh/examples/meshconn"
+	"github.com/weaveworks/mesh/meshconn"
 )
 
 func main() {
